devrequirements.txt
numpy==1.8.1
Pillow==2.5.1
matplotlib==1.4.1
coveralls